requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
openai>=1.40.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
streamlit==1.37.1
requests==2.31.0
pandas==2.2.2
tiktoken
